# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: MeshCrypt API
  description: Privacy-preserving wallet and bridge API with ZK-SNARKs
  version: 1.0.0
  contact:
    name: MeshCrypt Team
    email: support@meshcrypt.io
    url: https://meshcrypt.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.meshcrypt.io/v1
    description: Production server
  - url: https://api-staging.meshcrypt.io/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /transactions:
    post:
      summary: Create transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '202':
          description: Transaction accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /transactions/{txHash}:
    get:
      summary: Get transaction
      tags: [Transactions]
      parameters:
        - name: txHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /bridge/events:
    get:
      summary: Get bridge events
      tags: [Bridge]
      parameters:
        - name: chain_id
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Bridge events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/BridgeEvent'
                  count:
                    type: integer
components:
  schemas:
...
    TransactionRequest:
      type: object
      required: [wallet_id, to, amount]
      properties:
        wallet_id:
          type: string
        to:
          type: string
        amount:
          type: integer
          format: int64
        privacy_enabled:
          type: boolean
          default: true
    TransactionResponse:
      type: object
      properties:
        tx_hash:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
    Transaction:
      type: object
      properties:
        tx_hash:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          type: integer
          format: int64
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        privacy:
          $ref: '#/components/schemas/PrivacyData'
    PrivacyData:
      type: object
      properties:
        commitment:
          type: string
          format: byte
        range_proof:
          type: string
          format: byte
        stealth_address:
          type: string
        nullifier:
          type: string
          format: byte
...
    BridgeEvent:
      type: object
      properties:
        event_id:
          type: string
        chain_id:
          type: integer
        event_type:
          type: string
          enum: [deposit, withdraw]
        tx_hash:
          type: string
        timestamp:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKey: []